[objc] Add test for plain C calls that are defined inside obj-c headers
This test currently is cheating a little, since it adds a #ifdef __cplusplus
guard into the header itself. This is not reliable, since we depend that
any library we're linking adds these guards as well. A much better solution
would be to incorporate including plain C files into hxcpp itself.
Since this is actually a little independent from objective-c (even though
in the objective-c case, it would also be nice to use #import instead),
a new patch will be done later to propose a fix to this