[cpp] Add `-D scriptableReplace` that allows hxcpp-compiled classes to be replaced by cppia versions
This change requires hxcpp changes if `-D scriptableReplace` is defined. It generates a prelude
on static functions and on the constructor to make sure that the cppia loaded version takes precedence if it
is loaded.

Any change in source code is guarded against `-D scriptableReplace`, or its cpp define, `HXCPP_SCRIPTABLE_REPLACE`